+++ src/test/java/qna/domain/qna/QnATest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.qna;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
